bug: fixed seqnames mismatch bam and GATC data
I ran into this 'bug' when I processed new data from GCF. There is a
mismmismatch in names of non-canonical chromosomes (I don't understand
why).
The solution will be that I will limit the GATC.fragments data object
to only dedicated chromosomes (CHR: 1-22, X). This object will be
propagated to bam2fragcounts(...). I will use that to set ranges of
sequences for which data is retrieved from bamfiles. Only the last part
is fixed in this commit.

In addition to this commit I need to adapt the run-wrapper Rscript:
* restrict the GATC.mpbl.reads to 'CHR' chromosomes
* adapt the 'fExp' object to the same length as `GATC.mpbl.reads'